Module Ops.Unary

type t = string -> t

val register : string -> t -> unit

register name value register value with a given name. If name was already associated with some other value, then it will be superseded with the new binding.

val find : string -> t option

find name find a value associated with the given value

val list : unit -> (string * t) list

list () list all bindings